Move write operations to services
