+++ src/repositories/appSetting.repository.ts
@@ -34,37 +34,8 @@ export const findAllAppSettings = async ({
   };
 };
 
-export const createAppSetting = async (data: {
-  app_label: string;
-  app_type: string;
-  app_version: number;
-  force_updates: boolean;
-  maintenance_mode: boolean;
-}) => {
-  return prisma.appSetting.create({ data });
-};
-
-export const updateAppSetting = async (
-  id: number,
-  data: {
-    app_label: string;
-    app_type: string;
-    app_version: number;
-    force_updates: boolean;
-    maintenance_mode: boolean;
-  }
-) => {
-  return prisma.appSetting.update({
-    where: { id },
-    data,
+export const findAppSettingByType = async (app_type: string) => {
+  return prisma.appSetting.findFirst({
+    where: { app_type },
   });
-};
-
-export const softDeleteAppSetting = async (id: number): Promise<boolean> => {
-  await prisma.appSetting.update({
-    where: { id },
-    data: { deleted_at: new Date() },
-  });
-
-  return true;
 };
