+++ src/repositories/appVariable.repository.ts
@@ -6,17 +6,3 @@ export const getAppVariables = async () => {
     orderBy: { created_at: "desc" },
   });
 };
-
-export const createAppVariable = async (data: { name: string; value: string }) => {
-  return prisma.appVariable.create({ data });
-};
-
-export const updateAppVariable = async (
-  id: number,
-  data: { name: string; value: string }
-) => {
-  return prisma.appVariable.update({
-    where: { id },
-    data,
-  });
-};
