+++ src/services/appSetting.service.ts
@@ -1,0 +1,5 @@
+import { findAppSettingByType } from "../repositories/appSetting.repository";
+
+export const getAppSettingByType = async (appType: string) => {
+  return findAppSettingByType(appType);
+};
